Fixed circular import in builtins
